refactor: ajustes na estrutura de pacotes do projeto
