--- src/domain/entities/Cliente.java
@@ -1,4 +1,0 @@
-package domain.entities;
-
-public class Cliente {
-}
