+++ src/main/ecommerce/core/domain/entity/Cliente.java
@@ -1,0 +1,4 @@
+package main.ecommerce.core.domain.entity;
+
+public class Cliente {
+}
