+++ src/main/ecommerce/core/domain/interfaces/DescontoRepository.java
@@ -1,4 +1,4 @@
-package domain.interfaces;
+package main.ecommerce.core.domain.interfaces;
 
 public interface DescontoRepository {
     public double aplicarDesconto (double total);
