+++ src/main/ecommerce/core/domain/usecase/ProdutoUseCase.java
@@ -1,0 +1,4 @@
+package main.ecommerce.core.domain.usecase;
+
+public class ProdutoUseCase {
+}
