+++ src/main/ecommerce/global/CorsConfig.java
@@ -1,0 +1,4 @@
+package main.ecommerce.global;
+
+public class CorsConfig {
+}
