+++ src/main/ecommerce/infra/csv/Csv.java
@@ -1,0 +1,4 @@
+package main.ecommerce.infra.csv;
+
+public class Csv {
+}
