+++ src/main/ecommerce/infra/memory/Memory.java
@@ -1,0 +1,4 @@
+package main.ecommerce.infra.memory;
+
+public class Memory {
+}
